# actions/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Call Actions API",
    "version": "2.0.2"
  },
  "paths": {
    "/agents": {
      "get": {
        "summary": "List available agents (hierarchical)",
        "operationId": "agents",
        "x-openai-isConsequential": false,
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by project (supports * wildcard)",
              "default": "",
              "title": "Project"
            },
            "description": "Filter by project (supports * wildcard)"
          },
          {
            "name": "agent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by agent (supports * wildcard)",
              "default": "",
              "title": "Agent"
            },
            "description": "Filter by agent (supports * wildcard)"
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by prompt (supports * wildcard)",
              "default": "",
              "title": "Prompt"
            },
            "description": "Filter by prompt (supports * wildcard)"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/call": {
      "get": {
        "summary": "Call single agent/prompt via library",
        "operationId": "call",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The name used as 'target' for selection",
              "title": "Name"
            },
            "description": "The name used as 'target' for selection"
          },
          {
            "name": "input",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Input text",
              "title": "Input"
            },
            "description": "Input text"
          },
          {
            "name": "echo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, return structured JSON from library",
              "default": false,
              "title": "Echo"
            },
            "description": "If true, return structured JSON from library"
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Override session id: AgentName:chat or AgentName:chat:thread",
              "title": "Session Id"
            },
            "description": "Override session id: AgentName:chat or AgentName:chat:thread"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exec": {
      "post": {
        "summary": "Execute with a JSON payload (target|prompt|agent + context)",
        "operationId": "exec_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/read/{id}": {
      "get": {
        "summary": "Read raw card text from repo.db for any entity project/agent/prompt",
        "operationId": "read",
        "x-openai-isConsequential": false,
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/write/{id}": {
      "post": {
        "summary": "Write card text to repo.db and filesystem for any entity project/agent/prompt",
        "operationId": "write",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prompts": {
      "get": {
        "summary": "List prompts (ready/draft)",
        "operationId": "prompts",
        "x-openai-isConsequential": false,
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by project (exact)",
              "default": "",
              "title": "Project"
            },
            "description": "Filter by project (exact)"
          },
          {
            "name": "agent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by agent (exact)",
              "default": "",
              "title": "Agent"
            },
            "description": "Filter by agent (exact)"
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by prompt id or name (supports *)",
              "default": "",
              "title": "Prompt"
            },
            "description": "Filter by prompt id or name (supports *)"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by state: ready|draft|<empty for both>",
              "default": "",
              "title": "State"
            },
            "description": "Filter by state: ready|draft|<empty for both>"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExecPayload": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "context": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "echo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Echo",
            "default": true
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "title": "ExecPayload",
        "description": "Execute with a single selector. Exactly one of project|agent|prompt|target must be provided.",
        "oneOf": [
          {
            "required": [
              "project"
            ]
          },
          {
            "required": [
              "agent"
            ]
          },
          {
            "required": [
              "prompt"
            ]
          },
          {
            "required": [
              "target"
            ]
          }
        ],
        "x-exactly-one": [
          "project",
          "agent",
          "prompt",
          "target"
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "{{ CALL_ACTIONS_BASE_URL }}"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}